+++ pkg/test/storage_suite.go
@@ -9,30 +9,34 @@ import (
 	"testing"
 	"time"
 
+	"github.com/jhillyerd/inbucket/pkg/config"
 	"github.com/jhillyerd/inbucket/pkg/message"
 	"github.com/jhillyerd/inbucket/pkg/storage"
 )
 
 // StoreFactory returns a new store for the test suite.
-type StoreFactory func() (store storage.Store, destroy func(), err error)
+type StoreFactory func(config.Storage) (store storage.Store, destroy func(), err error)
 
 // StoreSuite runs a set of general tests on the provided Store.
 func StoreSuite(t *testing.T, factory StoreFactory) {
 	testCases := []struct {
 		name string
 		test func(*testing.T, storage.Store)
+		conf config.Storage
 	}{
-		{"metadata", testMetadata},
-		{"content", testContent},
-		{"delivery order", testDeliveryOrder},
-		{"size", testSize},
-		{"delete", testDelete},
-		{"purge", testPurge},
-		{"visit mailboxes", testVisitMailboxes},
+		{"metadata", testMetadata, config.Storage{}},
+		{"content", testContent, config.Storage{}},
+		{"delivery order", testDeliveryOrder, config.Storage{}},
+		{"size", testSize, config.Storage{}},
+		{"delete", testDelete, config.Storage{}},
+		{"purge", testPurge, config.Storage{}},
+		{"cap=10", testMsgCap, config.Storage{MailboxMsgCap: 10}},
+		{"cap=0", testNoMsgCap, config.Storage{MailboxMsgCap: 0}},
+		{"visit mailboxes", testVisitMailboxes, config.Storage{}},
 	}
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
-			store, destroy, err := factory()
+			store, destroy, err := factory(tc.conf)
 			if err != nil {
 				t.Fatal(err)
 			}
@@ -169,11 +173,11 @@ func testDeliveryOrder(t *testing.T, store storage.Store) {
 	subjects := []string{"alpha", "bravo", "charlie", "delta", "echo"}
 	for i, subj := range subjects {
 		// Check mailbox count.
-		getAndCountMessages(t, store, mailbox, i)
-		deliverMessage(t, store, mailbox, subj, time.Now())
+		GetAndCountMessages(t, store, mailbox, i)
+		DeliverToStore(t, store, mailbox, subj, time.Now())
 	}
 	// Confirm delivery order.
-	msgs := getAndCountMessages(t, store, mailbox, 5)
+	msgs := GetAndCountMessages(t, store, mailbox, 5)
 	for i, want := range subjects {
 		got := msgs[i].Subject()
 		if got != want {
@@ -189,7 +193,7 @@ func testSize(t *testing.T, store storage.Store) {
 	sentIds := make([]string, len(subjects))
 	sentSizes := make([]int64, len(subjects))
 	for i, subj := range subjects {
-		id, size := deliverMessage(t, store, mailbox, subj, time.Now())
+		id, size := DeliverToStore(t, store, mailbox, subj, time.Now())
 		sentIds[i] = id
 		sentSizes[i] = size
 	}
@@ -211,9 +215,9 @@ func testDelete(t *testing.T, store storage.Store) {
 	mailbox := "fred"
 	subjects := []string{"alpha", "bravo", "charlie", "delta", "echo"}
 	for _, subj := range subjects {
-		deliverMessage(t, store, mailbox, subj, time.Now())
+		DeliverToStore(t, store, mailbox, subj, time.Now())
 	}
-	msgs := getAndCountMessages(t, store, mailbox, len(subjects))
+	msgs := GetAndCountMessages(t, store, mailbox, len(subjects))
 	// Delete a couple messages.
 	err := store.RemoveMessage(mailbox, msgs[1].ID())
 	if err != nil {
@@ -225,7 +229,7 @@ func testDelete(t *testing.T, store storage.Store) {
 	}
 	// Confirm deletion.
 	subjects = []string{"alpha", "charlie", "echo"}
-	msgs = getAndCountMessages(t, store, mailbox, len(subjects))
+	msgs = GetAndCountMessages(t, store, mailbox, len(subjects))
 	for i, want := range subjects {
 		got := msgs[i].Subject()
 		if got != want {
@@ -233,9 +237,9 @@ func testDelete(t *testing.T, store storage.Store) {
 		}
 	}
 	// Try appending one more.
-	deliverMessage(t, store, mailbox, "foxtrot", time.Now())
+	DeliverToStore(t, store, mailbox, "foxtrot", time.Now())
 	subjects = []string{"alpha", "charlie", "echo", "foxtrot"}
-	msgs = getAndCountMessages(t, store, mailbox, len(subjects))
+	msgs = GetAndCountMessages(t, store, mailbox, len(subjects))
 	for i, want := range subjects {
 		got := msgs[i].Subject()
 		if got != want {
@@ -249,15 +253,52 @@ func testPurge(t *testing.T, store storage.Store) {
 	mailbox := "fred"
 	subjects := []string{"alpha", "bravo", "charlie", "delta", "echo"}
 	for _, subj := range subjects {
-		deliverMessage(t, store, mailbox, subj, time.Now())
+		DeliverToStore(t, store, mailbox, subj, time.Now())
 	}
-	getAndCountMessages(t, store, mailbox, len(subjects))
+	GetAndCountMessages(t, store, mailbox, len(subjects))
 	// Purge and verify.
 	err := store.PurgeMessages(mailbox)
 	if err != nil {
 		t.Fatal(err)
 	}
-	getAndCountMessages(t, store, mailbox, 0)
+	GetAndCountMessages(t, store, mailbox, 0)
+}
+
+// testMsgCap verifies the message cap is enforced.
+func testMsgCap(t *testing.T, store storage.Store) {
+	mbCap := 10
+	mailbox := "captain"
+	for i := 0; i < 20; i++ {
+		subj := fmt.Sprintf("subject %v", i)
+		DeliverToStore(t, store, mailbox, subj, time.Now())
+		msgs, err := store.GetMessages(mailbox)
+		if err != nil {
+			t.Fatalf("Failed to GetMessages for %q: %v", mailbox, err)
+		}
+		if len(msgs) > mbCap {
+			t.Errorf("Mailbox has %v messages, should be capped at %v", len(msgs), mbCap)
+			break
+		}
+		// Check that the first message is correct.
+		first := i - mbCap + 1
+		if first < 0 {
+			first = 0
+		}
+		firstSubj := fmt.Sprintf("subject %v", first)
+		if firstSubj != msgs[0].Subject() {
+			t.Errorf("Got subject %q, wanted first subject: %q", msgs[0].Subject(), firstSubj)
+		}
+	}
+}
+
+// testNoMsgCap verfies a cap of 0 is not enforced.
+func testNoMsgCap(t *testing.T, store storage.Store) {
+	mailbox := "captain"
+	for i := 0; i < 20; i++ {
+		subj := fmt.Sprintf("subject %v", i)
+		DeliverToStore(t, store, mailbox, subj, time.Now())
+		GetAndCountMessages(t, store, mailbox, i+1)
+	}
 }
 
 // testVisitMailboxes creates some mailboxes and confirms the VisitMailboxes method visits all of
@@ -265,8 +306,8 @@ func testPurge(t *testing.T, store storage.Store) {
 func testVisitMailboxes(t *testing.T, ds storage.Store) {
 	boxes := []string{"abby", "bill", "christa", "donald", "evelyn"}
 	for _, name := range boxes {
-		deliverMessage(t, ds, name, "Old Message", time.Now().Add(-24*time.Hour))
-		deliverMessage(t, ds, name, "New Message", time.Now())
+		DeliverToStore(t, ds, name, "Old Message", time.Now().Add(-24*time.Hour))
+		DeliverToStore(t, ds, name, "New Message", time.Now())
 	}
 	seen := 0
 	err := ds.VisitMailboxes(func(messages []storage.Message) bool {
@@ -285,9 +326,9 @@ func testVisitMailboxes(t *testing.T, ds storage.Store) {
 	}
 }
 
-// deliverMessage creates and delivers a message to the specific mailbox, returning the size of the
+// DeliverToStore creates and delivers a message to the specific mailbox, returning the size of the
 // generated message.
-func deliverMessage(
+func DeliverToStore(
 	t *testing.T,
 	store storage.Store,
 	mailbox string,
@@ -315,9 +356,9 @@ func deliverMessage(
 	return id, int64(len(testMsg))
 }
 
-// getAndCountMessages is a test helper that expects to receive count messages or fails the test, it
+// GetAndCountMessages is a test helper that expects to receive count messages or fails the test, it
 // also checks return error.
-func getAndCountMessages(t *testing.T, s storage.Store, mailbox string, count int) []storage.Message {
+func GetAndCountMessages(t *testing.T, s storage.Store, mailbox string, count int) []storage.Message {
 	t.Helper()
 	msgs, err := s.GetMessages(mailbox)
 	if err != nil {